requirements.txt
appdirs==1.4.4
Brotli==1.0.9
certifi==2023.5.7
click==8.1.3
mutagen==1.46.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pycryptodomex==3.17
pypdf==3.8.1
scenedetect==0.6.1
tqdm==4.65.0
websockets==11.0.3
yt-dlp==2023.3.4
